Refactor and enhance readability across multiple files
Refactored `CameraDetails.cs` to improve depth of field calculations and added a new method `CalculateDepthOfField`. Updated `JValueConverter.cs` to use `Int64` for `JTokenType.Integer`. Modified `Process.cs` to use a variable for `CodeType`. Refactored `RunController.cs` to improve test point handling and updated `ResultEntry` properties. Enhanced `DatabaseViewerViewModel.cs` to handle symbology types and close database connections properly. Reformatted `SymbologyMapView.xaml` and `TestViewModel.cs` for better readability and maintainability.

+++ ValueConverters/JValueConverter.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Globalization;
+using System.Numerics;
 using System.Windows.Data;
 using Newtonsoft.Json.Linq;
 
@@ -21,7 +22,7 @@ namespace JSONViewer_WPF.ValueConverters
                     case JTokenType.Bytes:
                         return (byte[])jval;
                     case JTokenType.Integer:
-                        return jval.Value<int>();
+                        return jval.Value<Int64>();
                 }
             }
